Fix warnings
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@46166 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/gtk3/joystickdrv/joystick_osx.c
@@ -132,8 +132,8 @@ static joystick_axis_t *make_axis(joy_hid_element_t *element, char *name)
 
     log_message(
         LOG_DEFAULT,
-        "joy-hid: axis: usage_page=0x%x usage=0x%x pmin=%ld pmax=%ld lmin=%ld lmax=%ld",
-        element->usage_page, element->usage, element->min_pvalue, element->max_pvalue, element->min_lvalue, element->max_lvalue);
+        "joy-hid: axis: usage_page=0x%x usage=0x%x pmin=%d pmax=%d lmin=%d lmax=%d",
+        (unsigned int)element->usage_page, (unsigned int)element->usage, element->min_pvalue, element->max_pvalue, element->min_lvalue, element->max_lvalue);
 
     axis = joystick_axis_new(name);
     axis->code = HID_CODE(element->usage_page, element->usage);
